Revert "feature: 支持密码重置可选择短信验证方式 #631 (#849)"
This reverts commit 7ef6234f9026bedc7133fab1cf5c70f12f46de73.
