+++ src/pages/src/common/demand-import.js
@@ -41,7 +41,6 @@ import {
   bkProgress,
   bkRadio,
   bkRadioGroup,
-  bkRadioButton,
   bkRoundProgress,
   bkRow,
   bkSearchSelect,
@@ -101,7 +100,6 @@ Vue.use(bkProcess);
 Vue.use(bkProgress);
 Vue.use(bkRadio);
 Vue.use(bkRadioGroup);
-Vue.use(bkRadioButton);
 Vue.use(bkRoundProgress);
 Vue.use(bkRow);
 Vue.use(bkSearchSelect);
