+++ src/pages/src/language/lang/en.js
@@ -174,13 +174,7 @@ export default {
   数字: 'number',
   '特殊字符（除空格）': 'Special characters (except spaces)',
   '请输入账户绑定的邮箱，我们将为您发送密码重置邮件': 'Please enter your account bound email address and we will send you a password reset email',
-  '请输入账号信息，我们将为您发送重置密码短信': 'Please enter your account information, and we will send you a message to reset your password',
-  '请输入用户名/手机号': 'Please enter username or mobile number',
-  发送验证码: 'Send verification code',
-  请输入验证码: 'Please enter the verification code',
   请输入邮箱: 'Please enter email address',
-  后: 'after',
-  重新发送: 'Resend',
   '邮箱格式错误，请重新输入': 'Mailbox format error, please re - enter',
   发送密码重置邮件: 'Send a password reset message',
   已发送密码重置邮件: 'A password reset message has been sent',
