+++ src/pages/src/language/lang/zh.js
@@ -174,12 +174,6 @@ export default {
   数字: '数字',
   '特殊字符（除空格）': '特殊字符（除空格）',
   '请输入账户绑定的邮箱，我们将为您发送密码重置邮件': '请输入账户绑定的邮箱，我们将为您发送密码重置邮件',
-  '请输入账号信息，我们将为您发送重置密码短信': '请输入账号信息，我们将为您发送重置密码短信',
-  '请输入用户名/手机号': '请输入用户名/手机号',
-  发送验证码: '发送验证码',
-  请输入验证码: '请输入验证码',
-  后: '后',
-  重新发送: '重新发送',
   请输入邮箱: '请输入邮箱',
   '邮箱格式错误，请重新输入': ' 邮箱格式错误，请重新输入',
   发送密码重置邮件: '发送密码重置邮件',
